Defect/build script filters (#600)
* Fixing Sitecore Filter to leverage Tag over Path

Some `Path` values contain .x.x (e.g. 9.x.x) instead of the full Sitecore Version value (e.g. `9.3.0`). As a result, they are not being included in the list of images to be built when using specifies their `SitecoreVersion` value.

* Update CHANGELOG.md

* Update Build.ps1

Converting change to an `-or` so as not to negatively impact build process intended for 10.x.